requirements.txt
mysqlclient==2.2.4
flask==3.0.3
mysql-connector==2.2.9
mysql-connector-python==9.0.0
docker==7.1.0
python-dotenv==1.0.1
tabulate
gunicorn
